refactor(folder+files): restructured folder+files to be easier to understand, added README's
