+++ lib/helpers/README.md
@@ -1,0 +1,5 @@
+# Helpers
+
+External helpers functions that are **NOT used by the modules code** itself, but are meant to be available for developers to import and use in their nuxtjs/firebase projects.
+
+See https://firebase.nuxtjs.org/guide/helpers .
