+++ lib/module.js
@@ -1,7 +1,7 @@
 const { resolve } = require('path')
 const firebase = require('firebase/app')
-const logger = require('./logger')
-const templateUtils = require('./template-utils')
+const logger = require('./utils/logger')
+const templateUtils = require('./utils/template-utils')
 
 const r = (...path) => resolve(__dirname, ...path)
 
@@ -84,7 +84,7 @@ module.exports = function (moduleOptions) {
 function addServiceWorker ({ config, onFirebaseHosting = false }, templateFile, templateOptions = {}) {
   // Add Service Worker Template
   this.addTemplate({
-    src: r(`templates/${templateFile}`),
+    src: r(`sw-templates/${templateFile}`),
     fileName: resolve(
       this.options.srcDir,
       this.options.dir.static,
