+++ lib/plugins/README.md
@@ -1,0 +1,6 @@
+# Plugins
+
+The files within the "plugins" folder will be parsed and copied to the /dist folder of the built project. These are then imported as nuxt-plugins.
+
+[Lodas Template](https://lodash.com/docs/4.17.15#template) is used to parse these plugins.
+
