+++ lib/plugins/services/app.js
@@ -1,14 +1,17 @@
 <%= options.writeStaticImports(options) %>
 
 export default async function createApp(config) {
-  <%= (stmt = options.writeImportStatement(options))
-  && `const firebase = ${stmt}`
+  <%= (statement = options.writeImportStatement(options))
+  && `const firebase = ${statement}`
   %>
 
   <% if (typeof options.sessions === 'object') { %>
+  /*************************************************
+  * If Firebase "Auth SSR serverLogin Option" is on
+  *************************************************/
   let session
   if (process.server && res && res.locals && res.locals.user) {
-    const { default: SessionManager } = await import('@nuxtjs/firebase/lib/services/SessionManager')
+    const { default: SessionManager } = await import('@nuxtjs/firebase/lib/utils/auth-ssr/ssr-auth-session-manager.js')
     const manager = new SessionManager(firebase, {
       config,
       sessions: <%= serialize(options.sessions) %>
@@ -20,7 +23,11 @@ export default async function createApp(config) {
   } else {
     session = firebase.apps.find(a => a.name === '[DEFAULT]') || firebase.initializeApp(config)
   }
+
   <% } else { %>
+  /*****************************************************
+  * Without "Auth SSR serverLogin Option" -> normal init
+  *****************************************************/
   if (!firebase.apps.length) {
     firebase.initializeApp(config)
   }
