+++ lib/sw-templates/README.md
@@ -1,0 +1,6 @@
+# SW-Templates
+
+These *service worker templates* will be parsed by lodash template and then copied in the /static folder of the built project.
+
+[Lodash Template](https://lodash.com/docs/4.17.15#template) is used to parse these plugins.
+
