+++ lib/sw-templates/firebase-auth-sw.js
