+++ lib/sw-templates/firebase-messaging-sw.js
